Move js/*.js -> views/utils && Join filters/*.js to util/filters.js
