--- frontend/simple/filters/index.js
@@ -1,1 +1,0 @@
-export { default as toPercent } from './toPercent.js'
