--- frontend/simple/filters/toPercent.js
@@ -1,1 +1,0 @@
-export default (decimal: number) => Math.floor(decimal * 100) + '%'
