+++ frontend/simple/views/components/CreateGroupSteps/GroupSummary.vue
@@ -106,8 +106,8 @@
   } */
 </style>
 <script>
-import { symbol } from '../../../js/currencies'
-import { toPercent } from '../../../filters'
+import { symbol } from '../../utils/currencies'
+import { toPercent } from '../../utils/filters'
 export default {
   name: 'GroupSummary',
   props: {
