+++ frontend/simple/views/components/ProgressOverview.vue
