+++ frontend/simple/views/utils/StepAssistant.js
