+++ frontend/simple/views/utils/currencies.js
