+++ frontend/simple/views/utils/filters.js
@@ -1,0 +1,1 @@
+export const toPercent = (decimal: number) => Math.floor(decimal * 100) + '%'
