+++ frontend/simple/views/views/Invite.vue
@@ -45,7 +45,7 @@
 <script>
 import * as Events from '../../../../shared/events'
 import backend from '../../controller/backend/'
-import L from '../../js/translations'
+import L from '../utils/translations'
 import template from 'string-template'
 import { GroupInvitees } from '../components/CreateGroupSteps'
 import { mapState, mapGetters } from 'vuex'
