+++ frontend/simple/views/views/Vote.vue
@@ -45,7 +45,7 @@
 import * as Events from '../../../../shared/events'
 import backend from '../../controller/backend/'
 import template from 'string-template'
-import L from '../../js/translations'
+import L from '../utils/translations'
 import _ from 'lodash'
 export default {
   name: 'Vote',
